+++ src/components/LocationSVG/index.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const LocationSVG = (props) => (
+  <svg className="LocationSVG"
+    {...props}>
+    <use xlinkHref="#LocationSymbol" />
+  </svg>
+);
+
+export default LocationSVG;