+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.2.0
+* [Main Section](https://github.com/ZorianaPro/hackathon2020/issues/4)
+* Add button component
+* Add Date and Location icons
+* Add Header component
+* Add SVGSpriteShit component
+* Add color variables
+
 ## 1.1.0
 * Deploy to GitHub pages